fix: Compile error by GCFloat in generic configuration (#235)
* fix: compile error by GCFloat

* build as generic

* add building with generic for tvOS

+++ WebDriverAgentLib/Utilities/FBActiveAppDetectionPoint.m
@@ -20,7 +20,7 @@
   if ((self = [super init])) {
     CGSize screenSize = [UIScreen mainScreen].bounds.size;
     // Consider the element, which is located close to the top left corner of the screen the on-screen one.
-    CGFloat pointDistance = MIN(screenSize.width, screenSize.height) * 0.2;
+    CGFloat pointDistance = MIN(screenSize.width, screenSize.height) * (CGFloat) 0.2;
     _coordinates = CGPointMake(pointDistance, pointDistance);
   }
   return self;
@@ -76,7 +76,7 @@
              withDescriptionFormat:@"Both screen point coordinates should be valid numbers. Got '%@' instead", coordinatesStr]
             buildError:error];
   }
-  self.coordinates = CGPointMake(strX.doubleValue, strY.doubleValue);
+  self.coordinates = CGPointMake((CGFloat) strX.doubleValue, (CGFloat) strY.doubleValue);
   return YES;
 }
